SLEMA-82 Separate date widgets
